FSM, my changes
